Fix sources.list when upgrading from old versions
In the section `Upgrading older releases` the `Get your current system fully updated` assumed that
the version being used contains all sources in `/etc/apt/sources.list`. Since 21.04 this isn't the case
as they are now stored in `/etc/apt/sources.list/*`.

This change will change the default line to use the new location and mention in a comment the 
old location as well.